made mouse control of pointer possible

+++ Game/BasicTank.pde
@@ -1,6 +1,6 @@
 public class BasicTank extends BubbleTank{
    
-  public BasicTank(float health,float radius, float id, float speed, float numBullets){
+  public BasicTank(float health,float radius, int id, float speed, float numBullets){
      super(health,radius,id,speed,numBullets);
   }
 }
